% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2Assign_function.R
\name{sc2Assign}
\alias{sc2Assign}
\alias{SC_sc2Assign}
\title{SC_sc2Assign Function}
\usage{
SC_sc2Assign(WAffinity_tabel, scData, percentile, reduction = "umap")

sc2Assign(
  count.matrix,
  norm.matrix,
  marker.genes,
  scData,
  percentile = 0.25,
  reduction = "umap"
)
}
\arguments{
\item{WAffinity_tabel}{A matrix of WAffinity scores for all cells.}

\item{scData}{A Seurat object containing cell metadata and dimensional reduction results.}

\item{percentile}{A numeric value specifying the threshold percentile for sc2Assign (default is 0.25).}

\item{reduction}{A string indicating the dimensionality reduction method to use (default is 'umap').}

\item{count.matrix}{A matrix containing gene expression counts where rows represent genes and columns represent cells.}

\item{norm.matrix}{A matrix containing normalized gene expression data for the same genes and cells.}

\item{marker.genes}{A list where each element is a vector of marker genes corresponding to a particular cell type.}
}
\value{
A list containing the misclassified cells and their reassigned cell types.

A list containing:
\describe{
  \item{WAffinity}{A vector of WAffinity-based cell type predictions.}
  \item{sc2Assign}{A vector of cell type predictions refined by the sc2Assign method.}
}
}
\description{
This function refines the WAffinity predictions by reassigning cells using the difference between the top two cell type scores.

This function performs cell type assignment using WAffinity and sc2Assign methods.
}
