% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2Assign_function.R
\name{sc2Assign_package}
\alias{sc2Assign_package}
\title{---
title: sc2Assign package}
\description{
---
title: sc2Assign package
}
\seealso{
See iScience 2024 paper.
}
\author{
M. Malmir, Y. Chen, Y. Jin
}
