% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2Assign_function.R
\name{SC_WAffinity}
\alias{SC_WAffinity}
\title{SC_WAffinity Function}
\usage{
SC_WAffinity(count.matrix, norm.matrix, marker.genes)
}
\arguments{
\item{count.matrix}{A matrix containing gene expression counts (rows = genes, columns = cells).}

\item{norm.matrix}{A matrix of normalized gene expression values for the same genes and cells.}

\item{marker.genes}{A list of marker genes for each cell type.}
}
\value{
A list containing WAffinity score table, predictions, and the gene weights used for each cell type.
}
\description{
This function computes the WAffinity scores for all cell types based on marker gene expression and weighted scores.
}
