% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2Assign_function.R
\name{WAffinity_Score}
\alias{WAffinity_Score}
\title{WAffinity_Score Function}
\usage{
WAffinity_Score(count.matrix, count.matrix.markers, weight)
}
\arguments{
\item{count.matrix}{A matrix of gene expression counts (rows = genes, columns = cells).}

\item{count.matrix.markers}{A matrix of gene expression counts for marker genes.}

\item{weight}{A numeric vector representing weights for each marker gene.}
}
\value{
A numeric vector of WAffinity scores for each cell.
}
\description{
This function calculates the WAffinity score for each cell based on marker gene expression.
}
