% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2Assign_function.R
\name{WAffinity_Weight}
\alias{WAffinity_Weight}
\title{WAffinity_Weight Function}
\usage{
WAffinity_Weight(norm.matrix.markers, gene.markers)
}
\arguments{
\item{norm.matrix.markers}{A normalized matrix of marker genes' expression data.}

\item{gene.markers}{A vector of marker genes.}
}
\value{
A numeric vector of gene scores (weights) for the marker genes.
}
\description{
This function calculates the WAffinity weights using the Pearson correlation coefficient between marker genes.
}
